requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.0
Flask-MySQLdb==2.0.0
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.3
mysqlclient==2.2.7
Werkzeug==3.1.3
zipp==3.20.2
dicttoxml==1.7.16
PyJWT==2.8.0
